# src/main/resources/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Swagger2 API",
    "version": "1.0",
    "title": "Spring Boot 后台API",
    "contact": {
      "name": "frank",
      "url": "https://barry215.github.io",
      "email": "frank315111@gmail.com"
    }
  },
  "host": "localhost:8888",
  "basePath": "/",
...
  "paths": {
...
    "/p/archive/count": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取归档数",
        "description": "获取归档数",
        "operationId": "getArchiveArticlesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/archive/{offset}/{size}": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取归档文章",
        "description": "获取归档文章",
        "operationId": "getArchiveArticlesUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "path",
            "description": "页码",
            "required": true,
            "type": "ref"
          },
          {
            "name": "size",
            "in": "path",
            "description": "每页数量",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/archives": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取归档数条目",
        "description": "获取归档条目",
        "operationId": "getArchivesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/archives/{archive}/u": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取归档内文章",
        "description": "获取归档内文章",
        "operationId": "getArticleInArchivesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "archive",
            "in": "path",
            "description": "archive",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/count": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取文章总数",
        "description": "获取文章总数",
        "operationId": "getArticleCountUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/edit": {
      "put": {
        "tags": [
          "blog-controller"
        ],
        "summary": "修改文章",
        "description": "修改文章",
        "operationId": "modifyArticleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "articleModifyForm",
            "description": "表单",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ArticleModifyForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/edit/u": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取草稿列表",
        "description": "获取草稿列表",
        "operationId": "getEditArticlesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/edit/{document_id}": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取草稿文章",
        "description": "获取草稿文章",
        "operationId": "getEditArticlesUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "description": "文章ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/h/{document_id}": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取历史文章",
        "description": "获取历史文章",
        "operationId": "getHistoryArticlesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "description": "文章ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/new": {
      "post": {
        "tags": [
          "blog-controller"
        ],
        "summary": "创建文章",
        "description": "创建文章",
        "operationId": "createArticleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "articleForm",
            "description": "示例实体",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ArticleForm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/tag/count": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取标签数",
        "description": "获取标签数",
        "operationId": "getTagsCountUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/u/{offset}/{size}": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取文章列表",
        "description": "获取文章列表",
        "operationId": "getDocumentsListUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "path",
            "description": "页码",
            "required": true,
            "type": "ref"
          },
          {
            "name": "size",
            "in": "path",
            "description": "每页数量",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/{document_id}": {
      "get": {
        "tags": [
          "blog-controller"
        ],
        "summary": "获取文章",
        "description": "获取文章",
        "operationId": "getArticleUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "description": "文章ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/p/{document_id}/check/{version}": {
      "put": {
        "tags": [
          "blog-controller"
        ],
        "summary": "切换文章版本",
        "description": "切换文章版本",
        "operationId": "checkVersionUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "description": "文档ID",
            "required": true,
            "type": "ref"
          },
          {
            "name": "version",
            "in": "path",
            "description": "版本号",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JsonResult«object»"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "ArticleForm": {
      "type": "object",
      "properties": {
        "archive": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "hasPublished": {
          "type": "integer",
          "format": "int32"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
    "ArticleModifyForm": {
      "type": "object",
      "properties": {
        "archive": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "document_id": {
          "type": "integer",
          "format": "int32"
        },
        "hasPublished": {
          "type": "integer",
          "format": "int32"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "title": {
          "type": "string"
        },
        "version": {
          "type": "string"
        }
      }
    },
...
    "JsonResult«object»": {
      "type": "object",
      "properties": {
        "data": {
          "type": "object"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  }
}